YAY-114 refactor: transfer styles, store to app
