+++ src/app/globals.css
@@ -1,4 +1,4 @@
-@import '../styles/index.css';
+@import './styles/index.css';
 
 @import 'tailwindcss';
