+++ src/app/layout.tsx
@@ -1,5 +1,5 @@
 import { Alert } from '@/shared/components/Alert'
-import { StoreProvider } from '@/shared/store/providers'
+import { StoreProvider } from '@/app/store/providers'
 import { Header } from '@/widgets/Header'
 import type { Metadata } from 'next'
 import { Inter } from 'next/font/google'
