+++ src/app/store/base-api.ts
@@ -1,6 +1,6 @@
 'use client'
 import { createApi } from '@reduxjs/toolkit/query/react'
-import { baseQueryWithReAuth } from '@/shared/store/base-query'
+import { baseQueryWithReAuth } from '@/app/store/base-query'
 
 export const baseApi = createApi({
    reducerPath: 'baseApi',
