+++ src/entities/user/api/user.service.ts
@@ -1,5 +1,5 @@
 'use client'
-import { baseApi } from '@/shared/store'
+import { baseApi } from '@/app/store'
 import { UserEndpoints } from '@/shared/enums'
 import { UserProfileType } from '@/entities/user/api'
