+++ src/features/settings/api/devices.service.ts
@@ -1,5 +1,5 @@
 'use client'
-import { baseApi } from '@/shared/store'
+import { baseApi } from '@/app/store'
 import { getDevicesResponse } from './settings.types'
 import { UserEndpoints } from '@/shared/enums'
