+++ src/features/settings/api/payments.service.ts
@@ -1,5 +1,5 @@
 'use client'
-import { baseApi } from '@/shared/store'
+import { baseApi } from '@/app/store'
 import {
    CreateSubscriptionArgs,
    CreateSubscriptionResponse,
