+++ src/shared/hooks/useAppDispatch.ts
@@ -1,4 +1,4 @@
 import { useDispatch } from 'react-redux'
-import { AppDispatch } from '../store'
+import { AppDispatch } from '../../app/store'
 
 export const useAppDispatch = () => useDispatch<AppDispatch>()
