+++ src/shared/hooks/useAuthSync.ts
@@ -6,7 +6,7 @@ import { authChannel } from '@/shared/lib/authBroadcast'
 import { useAppDispatch } from '@/shared/hooks/useAppDispatch'
 import { PrivateRoutes, PublicRoutes } from '@/shared/enums'
 import { authService } from '@/features/auth/api'
-import { baseApi } from '@/shared/store'
+import { baseApi } from '@/app/store'
 
 export const useAuthSync = () => {
    const router = useRouter()
